+++ core/src/main/scala/org/http4s/multipart/MultipartParser.scala
@@ -20,7 +20,7 @@ object MultipartParser {
   private final case class Out[+A](a: A, tail: Option[ByteVector] = None)
 
   def parse(boundary: Boundary): Writer1[Headers, ByteVector, ByteVector] =
-    parse(boundary, 10 * 1024)
+    parse(boundary, 40 * 1024)
 
   def parse(boundary: Boundary, headerLimit: Long): Writer1[Headers, ByteVector, ByteVector] = {
     val boundaryBytes = boundary.toByteVector
@@ -32,7 +32,7 @@ object MultipartParser {
 
     def receiveLine(leading: Option[ByteVector]): Writer1[ByteVector, ByteVector, Out[ByteVector]] = {
       def handle(bv: ByteVector): Writer1[ByteVector, ByteVector, Out[ByteVector]] = {
-        logger.trace(s"handle: $bv")
+        logger.trace(s"handle: ${bv.decodeUtf8}")
         val index = bv indexOfSlice CRLF
 
         if (index >= 0) {
@@ -54,9 +54,9 @@ object MultipartParser {
       leading map handle getOrElse receive1(handle)
     }
 
-    def takeUpTo(n: Long, pf: => ParseFailure): Process1[ByteVector, ByteVector] =
-      if (n > 0) receive1(bv => emit(bv) ++ takeUpTo(n - bv.size, pf))
-      else fail(pf)
+    def takeUpTo(n: Long): Process1[ByteVector, ByteVector] =
+      if (n > 0) receive1(bv => emit(bv) ++ takeUpTo(n - bv.size))
+      else fail(MalformedMessageBodyFailure(s"Part header was longer than ${headerLimit}-byte limit"))
 
     def receiveCollapsedLine(leading: Option[ByteVector]): Process1[ByteVector, Out[ByteVector]] = {
       receiveLine(leading) pipe process1.fold(Out(ByteVector.empty)) {
@@ -66,21 +66,13 @@ object MultipartParser {
     }
 
     def start: Writer1[Headers, ByteVector, ByteVector] = {
-      val partHeaderTooLong =
-        ParseFailure("Invalid multipart entity", s"Part header was longer than ${headerLimit}-byte limit")
-
-      logger.info("Exepected " + expected)
-
       def preamble(leading: Option[ByteVector]): Process1[ByteVector, Option[ByteVector]] =
         body(leading.map(_.compact), expected).flatMap {
           case Out(chunk, None) =>
-            logger.debug(s"preamble chunk: $chunk")
+            logger.trace(s"preamble chunk: ${chunk.decodeUtf8}")
             halt
-          case Out(ByteVector.empty, tail) =>
-            logger.trace(s"Resuming preamble with $tail.")
-            emit(tail)
           case Out(chunk, tail) =>
-            logger.debug(s"Last preamble chunk: $chunk.")
+            logger.trace(s"Last preamble chunk: $chunk.")
             logger.trace(s"Resuming with $tail.")
             emit(tail)
         }
@@ -90,7 +82,7 @@ object MultipartParser {
           line.startsWith(startLine) && isTransportPadding(line.drop(startLine.size))
 
         def isEndLine(line: ByteVector): Boolean =
-          line.startsWith(endLine) && isTransportPadding(line.drop(endLine.size))
+          line.startsWith(endLine)
 
         def isTransportPadding(bv: ByteVector): Boolean =
           bv.toSeq.find(b => b != ' ' && b != '\t').isEmpty
@@ -101,28 +93,20 @@ object MultipartParser {
             emit(tail)
           }
           else if (isEndLine(line)) {
-            logger.debug("Found end line. Halting.")
-            halt
+            logger.debug("Found end line. Discarding epilogue.")
+            process1.skip.repeat
           }
-          else if (line.nonEmpty) {
-            logger.trace(s"Discarding preamble: $line")
-            beginPart(tail)
-          }
-          else {
-            halt
-          }
+          else fail(new MalformedMessageBodyFailure(s"Expected a multipart start or end line: ${line.decodeUtf8}"))
         }
       }
 
       def headers(leading: Option[ByteVector]): Process1[ByteVector, Out[Headers]] = {
-        takeUpTo(headerLimit, partHeaderTooLong) pipe (for {
+        takeUpTo(headerLimit) pipe (for {
           tail <- beginPart(leading)
           headerPair <- header(tail, expected)
           Out(headers, tail2) = headerPair
           _ = logger.debug(s"Headers: $headers")
-          spacePair <- receiveCollapsedLine(tail2)
-          tail3 = spacePair.tail // eat the space between header and content
-        } yield Out(headers, tail3))
+        } yield Out(headers, tail2))
       }
 
       def go(leading: Option[ByteVector]): Writer1[Headers, ByteVector, ByteVector] = {
@@ -144,38 +128,46 @@ object MultipartParser {
         } yield part
       }
 
-      // RFC1341 says we can tolerate a multipart body that begins
-      // with an enacpsulation line NOT initiated by a CRLF.  If we
-      // supply our own to start the preamble, we are tolerant.  If
-      // it's in the body, it gets eaten as part of the preamble.
+      // The requirement that the encapsulation boundary begins with a CRLF
+      // implies that the body of a multipart entity must itself begin with a
+      // CRLF before the first encapsulation line - that is, if the "preamble"
+      // area is not used, the entity headers must be followed by TWO CRLFs.
+      // This is indeed how such entities should be composed.  A tolerant mail
+      // reading program, however, may interpret a body of type multipart that
+      // begins with an encapsulation line NOT initiated by a CRLF as also
+      // being an encapsulation boundary, but a compliant mail sending program
+      // must not generate such entities. -- https://tools.ietf.org/html/rfc1341
+      //
+      // If we supply our own to start the preamble via `Some(CRLF)`,
+      // we are tolerant.  If it's in the body, it gets eaten as part
+      // of the preamble.
       preamble(Some(CRLF)).flatMap(go)
     }
 
     def header(leading: Option[ByteVector], expected: ByteVector): Process1[ByteVector, Out[Headers]] = {
-      def go(leading: Option[ByteVector], expected: ByteVector): Process1[ByteVector, Out[Headers]] = {
+      def go(leading: Option[ByteVector]): Writer1[Header, ByteVector, Option[ByteVector]] = {
+        logger.trace(s"Header go: ${leading.map(_.decodeUtf8)}  ")
         receiveCollapsedLine(leading) flatMap {
           case Out(bv, tail) => {
             if (bv == expected) {
-              halt
+              emitO(tail)
             } else {
-              val headerM = for {
+              (for {
                 line <- bv.decodeAscii.right.toOption
                 idx <- Some(line indexOf ':')
                 if idx >= 0
                 if idx < line.length - 1
-              } yield Header(line.substring(0, idx), line.substring(idx + 1).trim)
-
-              headerM.map { header => emit(Out(Headers(header), tail)) }
-                .map { _ ++ go(tail, expected) }
-                .getOrElse(halt)
+                header = Header(line.substring(0, idx), line.substring(idx + 1).trim)
+              } yield emitW(header) ++ go(tail)).getOrElse(emitO(tail))
             }
           }
         }
       }
 
-      go(leading, expected).fold(Out(Headers.empty)) {
-        case (acc, Out(header, tail)) => Out(acc.a ++ header, tail)
-      }
+      go(leading).map(x => {logger.trace("Got: "+x.toString); x}).fold(Out(List.empty[Header])) {
+        case (acc, -\/(header)) => acc.copy(a = header :: acc.a)
+        case (acc, \/-(tail)) => acc.copy(tail = tail)
+      }.map { case Out(hs, tail) => Out(Headers(hs.reverse), tail) }
     }
 
     def body(leading: Option[ByteVector], expected: ByteVector): Process1[ByteVector, Out[ByteVector]] = {